05_vcd_raster/pyproject.toml
[project]
name = "05-vcd-raster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pillow"]
